Zulip-terminal: Restructure code into separate files for each of MVC.
Create separate files for Model, View and Controller classes.
