+++ helper.py
@@ -1,0 +1,14 @@
+from typing import Any
+from functools import wraps
+from threading import Thread
+
+def async(func: Any) -> Any:
+    """
+    Decorator for executing a function in a separate :class:`threading.Thread`.
+    """
+    @wraps(func)
+    def wrapper(*args, **kwargs):
+        thread = Thread(target=func, args=args, kwargs=kwargs)
+        thread.daemon = True
+        return thread.start()
+    return wrapper
